rework project according to hexagonal architecture
