+++ internal/models/crawler.go
@@ -1,0 +1,7 @@
+package models
+
+type CrawlResult struct {
+	Err   error
+	Title string
+	Url   string
+}
